+++ src/encoding/general.rs
@@ -52,20 +52,24 @@ delegate_value_encoding!(delegate from (General) to (Map<General, General>)
     with generics (K, V));
 #[cfg(feature = "std")]
 delegate_encoding!(delegate from (General) to (Unpacked<General>)
-    for type (std::collections::HashSet<T>) with generics (T));
+    for type (std::collections::HashSet<T, S>)
+    with where clause (S: Default + core::hash::BuildHasher)
+    with generics (T, S));
 #[cfg(feature = "std")]
 delegate_value_encoding!(delegate from (General) to (Map<General, General>)
-    for type (std::collections::HashMap<K, V>)
-    with where clause (K: Eq + core::hash::Hash)
-    with generics (K, V));
+    for type (std::collections::HashMap<K, V, S>)
+    with where clause (K: Eq + core::hash::Hash, S: Default + core::hash::BuildHasher)
+    with generics (K, V, S));
 #[cfg(feature = "hashbrown")]
 delegate_encoding!(delegate from (General) to (Unpacked<General>)
-    for type (hashbrown::HashSet<T>) with generics (T));
+    for type (hashbrown::HashSet<T, S>)
+    with where clause (S: Default + core::hash::BuildHasher)
+    with generics (T, S));
 #[cfg(feature = "hashbrown")]
 delegate_value_encoding!(delegate from (General) to (Map<General, General>)
-    for type (hashbrown::HashMap<K, V>)
-    with where clause (K: Eq + core::hash::Hash)
-    with generics (K, V));
+    for type (hashbrown::HashMap<K, V, S>)
+    with where clause (K: Eq + core::hash::Hash, S: Default + core::hash::BuildHasher)
+    with generics (K, V, S));
 
 // General encodes bool and integers as varints.
 delegate_value_encoding!(delegate from (General) to (Varint)
@@ -174,14 +178,13 @@ impl ValueEncoder<General> for String {
 
 impl DistinguishedValueEncoder<General> for String {
     #[inline]
-    fn decode_value_distinguished<B: Buf + ?Sized>(
+    fn decode_value_distinguished<const ALLOW_EMPTY: bool>(
         value: &mut String,
-        buf: Capped<B>,
-        allow_empty: bool,
+        buf: Capped<impl Buf + ?Sized>,
         ctx: DecodeContext,
     ) -> Result<Canonicity, DecodeError> {
         Self::decode_value(value, buf, ctx)?;
-        Ok(if !allow_empty && value.is_empty() {
+        Ok(if !ALLOW_EMPTY && value.is_empty() {
             Canonicity::NotCanonical
         } else {
             Canonicity::Canonical
@@ -255,16 +258,14 @@ impl ValueEncoder<General> for Cow<'_, str> {
 
 impl DistinguishedValueEncoder<General> for Cow<'_, str> {
     #[inline]
-    fn decode_value_distinguished<B: Buf + ?Sized>(
+    fn decode_value_distinguished<const ALLOW_EMPTY: bool>(
         value: &mut Cow<str>,
-        buf: Capped<B>,
-        allow_empty: bool,
+        buf: Capped<impl Buf + ?Sized>,
         ctx: DecodeContext,
     ) -> Result<Canonicity, DecodeError> {
-        DistinguishedValueEncoder::<General>::decode_value_distinguished(
+        DistinguishedValueEncoder::<General>::decode_value_distinguished::<ALLOW_EMPTY>(
             value.to_mut(),
             buf,
-            allow_empty,
             ctx,
         )
     }
@@ -337,14 +338,13 @@ impl ValueEncoder<General> for bytestring::ByteString {
 #[cfg(feature = "bytestring")]
 impl DistinguishedValueEncoder<General> for bytestring::ByteString {
     #[inline]
-    fn decode_value_distinguished<B: Buf + ?Sized>(
+    fn decode_value_distinguished<const ALLOW_EMPTY: bool>(
         value: &mut bytestring::ByteString,
-        buf: Capped<B>,
-        allow_empty: bool,
+        buf: Capped<impl Buf + ?Sized>,
         ctx: DecodeContext,
     ) -> Result<Canonicity, DecodeError> {
         Self::decode_value(value, buf, ctx)?;
-        Ok(if !allow_empty && value.is_empty() {
+        Ok(if !ALLOW_EMPTY && value.is_empty() {
             Canonicity::NotCanonical
         } else {
             Canonicity::Canonical
@@ -416,14 +416,13 @@ impl ValueEncoder<General> for Bytes {
 
 impl DistinguishedValueEncoder<General> for Bytes {
     #[inline]
-    fn decode_value_distinguished<B: Buf + ?Sized>(
+    fn decode_value_distinguished<const ALLOW_EMPTY: bool>(
         value: &mut Bytes,
-        buf: Capped<B>,
-        allow_empty: bool,
+        buf: Capped<impl Buf + ?Sized>,
         ctx: DecodeContext,
     ) -> Result<Canonicity, DecodeError> {
         Self::decode_value(value, buf, ctx)?;
-        Ok(if !allow_empty && value.is_empty() {
+        Ok(if !ALLOW_EMPTY && value.is_empty() {
             Canonicity::NotCanonical
         } else {
             Canonicity::Canonical
@@ -472,16 +471,14 @@ impl ValueEncoder<General> for Blob {
 
 impl DistinguishedValueEncoder<General> for Blob {
     #[inline]
-    fn decode_value_distinguished<B: Buf + ?Sized>(
+    fn decode_value_distinguished<const ALLOW_EMPTY: bool>(
         value: &mut Blob,
-        buf: Capped<B>,
-        allow_empty: bool,
+        buf: Capped<impl Buf + ?Sized>,
         ctx: DecodeContext,
     ) -> Result<Canonicity, DecodeError> {
-        DistinguishedValueEncoder::<PlainBytes>::decode_value_distinguished(
+        DistinguishedValueEncoder::<PlainBytes>::decode_value_distinguished::<ALLOW_EMPTY>(
             &mut **value,
             buf,
-            allow_empty,
             ctx,
         )
     }
@@ -541,10 +538,9 @@ where
     T: RawDistinguishedMessage + Eq,
 {
     #[inline]
-    fn decode_value_distinguished<B: Buf + ?Sized>(
+    fn decode_value_distinguished<const ALLOW_EMPTY: bool>(
         value: &mut T,
-        mut buf: Capped<B>,
-        allow_empty: bool,
+        mut buf: Capped<impl Buf + ?Sized>,
         ctx: DecodeContext,
     ) -> Result<Canonicity, DecodeError> {
         ctx.limit_reached()?;
@@ -552,7 +548,7 @@ where
         // Empty message types always encode and decode from zero bytes. It is far cheaper to check
         // here than to check after the value has been decoded and checking the message's
         // `is_empty()`.
-        if !allow_empty && buf.remaining_before_cap() == 0 {
+        if !ALLOW_EMPTY && buf.remaining_before_cap() == 0 {
             return Ok(Canonicity::NotCanonical);
         }
         merge_distinguished(value, buf, ctx.enter_recursion())